+++ projects/packages/blaze/src/class-dashboard-rest-controller.php
@@ -96,6 +96,17 @@ class Dashboard_REST_Controller {
 			)
 		);
 
+		// WordAds DSP API upload to WP Media Library routes
+		register_rest_route(
+			static::$namespace,
+			sprintf( '/sites/%1$d/wordads/dsp/api/v1/wpcom/sites/%1$d/media', $site_id ),
+			array(
+				'methods'             => WP_REST_Server::CREATABLE,
+				'callback'            => array( $this, 'upload_image_to_current_website' ),
+				'permission_callback' => array( $this, 'can_user_view_dsp_callback' ),
+			)
+		);
+
 		// WordAds DSP API media openverse query routes
 		register_rest_route(
 			static::$namespace,
@@ -383,6 +394,65 @@ class Dashboard_REST_Controller {
 			return array();
 		}
 		return $this->get_dsp_generic( sprintf( 'v1/wpcom/sites/%d/media', $site_id ), $req );
+	}
+
+	/**
+	 * Redirect POST requests to WordAds DSP Blaze media endpoint for the site.
+	 *
+	 * @return array|WP_Error
+	 */
+	public function upload_image_to_current_website() {
+		$site_id = $this->get_site_id();
+		if ( is_wp_error( $site_id ) ) {
+			return array( 'error' => $site_id->get_error_message() );
+		}
+
+		if ( empty( $_FILES['image'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Missing
+			return array( 'error' => 'File is missed' );
+		}
+		$file      = $_FILES['image']; // phpcs:ignore WordPress.Security.NonceVerification.Missing,WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+		$temp_name = $file['tmp_name'] ?? ''; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+		if ( ! $temp_name || ! is_uploaded_file( $temp_name ) ) {
+			return array( 'error' => 'Specified file was not uploaded' );
+		}
+
+		// Getting the original file name.
+		$filename = sanitize_file_name( basename( $file['name'] ) );
+		// Upload contents to the Upload folder locally.
+		$upload = wp_upload_bits(
+			$filename,
+			null,
+			file_get_contents( $temp_name ) // phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents
+		);
+
+		if ( ! empty( $upload['error'] ) ) {
+			return array( 'error' => $upload['error'] );
+		}
+
+		// Check the type of file. We'll use this as the 'post_mime_type'.
+		$filetype = wp_check_filetype( $filename, null );
+
+		// Prepare an array of post data for the attachment.
+		$attachment = array(
+			'guid'           => wp_upload_dir()['url'] . '/' . $filename,
+			'post_mime_type' => $filetype['type'],
+			'post_title'     => preg_replace( '/\.[^.]+$/', '', $filename ),
+			'post_content'   => '',
+			'post_status'    => 'inherit',
+		);
+
+		// Insert the attachment.
+		$attach_id = wp_insert_attachment( $attachment, $upload['file'] );
+
+		// Make sure wp_generate_attachment_metadata() has all requirement dependencies.
+		require_once ABSPATH . 'wp-admin/includes/image.php';
+
+		// Generate the metadata for the attachment, and update the database record.
+		$attach_data = wp_generate_attachment_metadata( $attach_id, $upload['file'] );
+		// Store metadata in the local DB.
+		wp_update_attachment_metadata( $attach_id, $attach_data );
+
+		return array( 'url' => $upload['url'] );
 	}
 
 	/**