+++ projects/plugins/jetpack/modules/masterbar/nudges/additional-css/class-atomic-additional-css-manager.php
@@ -39,16 +39,7 @@ class Atomic_Additional_CSS_Manager {
 	 */
 	public function register_nudge( \WP_Customize_Manager $wp_customize_manager ) {
 		$nudge_url  = $this->get_nudge_url();
-		$nudge_text = __( 'Purchase a Business Plan to<br> activate CSS customization', 'jetpack' );
-
-		if (
-			( defined( 'ENABLE_PRO_PLAN' ) && ENABLE_PRO_PLAN ) ||
-			! empty( $_GET['enable_pro_plan'] ) || // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- Only used to customize output.
-			! empty( $_COOKIE['enable_pro_plan'] )
-		) {
-			$nudge_text = __( 'Purchase a Pro Plan to<br> activate CSS customization', 'jetpack' );
-			$nudge_url  = preg_replace( '/premium$/', 'pro', $nudge_url );
-		}
+		$nudge_text = __( 'Purchase the Creator plan to<br> activate CSS customization', 'jetpack' );
 
 		$nudge = new CSS_Customizer_Nudge(
 			$nudge_url,