+++ projects/js-packages/ai-client/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2024-02-13
+### Changed
+- Updated package dependencies. [#35608]
+
+## [0.6.0] - 2024-02-05
+### Added
+- Jetpack AI: Support floating error messaging on the AI Control component. [#35322]
+
+### Changed
+- Updated package dependencies. [#35384]
+
 ## [0.5.1] - 2024-01-29
 ### Changed
 - Update dependencies. [#35170]
@@ -199,6 +210,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated package dependencies. [#31659]
 - Updated package dependencies. [#31785]
 
+[0.6.1]: https://github.com/Automattic/jetpack-ai-client/compare/v0.6.0...v0.6.1
+[0.6.0]: https://github.com/Automattic/jetpack-ai-client/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/Automattic/jetpack-ai-client/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/Automattic/jetpack-ai-client/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/Automattic/jetpack-ai-client/compare/v0.4.0...v0.4.1