+++ projects/js-packages/licensing/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.12.1 - 2024-02-13
+### Changed
+- Updated package dependencies. [#35608]
+
+## 0.12.0 - 2024-02-05
+### Added
+- Added social specific success card on Jetpack Social license activation. [#35224]
+
+### Changed
+- Updated package dependencies.
+
 ## 0.11.20 - 2024-01-29
 ### Changed
 - Update dependencies.