+++ projects/js-packages/partner-coupon/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.69 - 2024-02-13
+### Changed
+- Updated package dependencies. [#35608]
+
+## 0.2.68 - 2024-02-05
+### Changed
+- Updated package dependencies.
+
 ## 0.2.67 - 2024-01-29
 ### Changed
 - Update dependencies. [#35170]