+++ projects/packages/heartbeat/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2024-02-07
+### Changed
+- The package now requires PHP >= 7.0. [#34192]
+
 ## [1.5.3] - 2023-07-06
 ### Added
 - Add Jetpack Autoloader package suggestion. [#29988]
@@ -124,6 +128,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use new heartbeat package
 - Creates the Jetpack Heartbeat package
 
+[2.0.0]: https://github.com/Automattic/jetpack-heartbeat/compare/v1.5.3...v2.0.0
 [1.5.3]: https://github.com/Automattic/jetpack-heartbeat/compare/v1.5.2...v1.5.3
 [1.5.2]: https://github.com/Automattic/jetpack-heartbeat/compare/v1.5.1...v1.5.2
 [1.5.1]: https://github.com/Automattic/jetpack-heartbeat/compare/v1.5.0...v1.5.1